+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Features**:
 
 - Add error and sample rate fields to the replay event parser. ([#1745](https://github.com/getsentry/relay/pull/1745))
+- Add `source` field to sdk interface in event payload. ([#1760](https://github.com/getsentry/relay/pull/1760))
 
 ## 23.1.0