+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Features**:
 
 - Add error and sample rate fields to the replay event parser. ([#1745](https://github.com/getsentry/relay/pull/1745))
+- Add SSL support to `relay-redis` crate. ([#1772](https://github.com/getsentry/relay/pull/1772))
 
 **Internal**: